pyproject.toml
[project]
name = "revenuecat-agent"
version = "0.1.0"
description = "AI Developer Advocate agent for RevenueCat"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.2.0",
    "supabase>=2.0.0",
    "pgvector>=0.3.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "firecrawl-py>=1.0.0",
    "e2b-code-interpreter>=1.0.0",
    "tweepy>=4.14.0",
    "streamlit>=1.40.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agents*", "tools*", "dashboard*", "scripts*", "database*", "config*"]
